[finance_report_vision] Split the last two >1000-line modules into packages
Splits the final two oversized hand-written modules (after reporting #1289,
extraction #1297, ai_advisor #1299, governance_report #1302, market_data #1310).
Both are behaviour-preserving; public surfaces preserved via __init__ re-exports.

## pr_preview_lifecycle.py (1623 -> 6 files)
tools/_lib/dev/pr_preview_lifecycle/{_base,_util,_preview,_dokploy,cli}.py.
Public surface (main / DokployConfig / dokploy_api_call) preserved; the
tools/pr_preview_lifecycle.py wrapper (invoked by the CI workflows) still imports
main. Test decoupling: tests patch ~18 core helpers (run_command,
dokploy_api_call, the compose CRUD) with setattr(lifecycle, "X") — single AND
multi-line forms. Fix: full re-export + repoint each patch to the submodule that
defines X (multi-line-aware), and module-qualify the few cross-module calls
(cli -> _dokploy, _dokploy -> _util) so the single repoint target stays
consistent. `import time` kept in __init__ for the lifecycle.time.* patches.
The source-scanning guards (test_post_merge_e2e_gates / test_cleanup_pr_preview_
resources) are made dir-aware and pointed at the package.

## lint_doc_consistency.py (1018 -> 6 files)
common/ssot/lint_doc_consistency/{_base,_types,_parsing,_checks,cli}.py.
Public surface (main) preserved; the tools/lint_doc_consistency.py wrapper still
imports it. Test decoupling: tests patch module CONSTANTS via
mock.patch.object(ldc, "REPO_ROOT"/"VISION_PATH"/"EPIC_DIR"/"TEST_ROOTS"/
"AC_REGISTRY"/"INFRA_REGISTRY"/"THIN_README_LIMITS"). Fix: full re-export +
module-qualify each patched constant at its cross-module use sites
(_base.REPO_ROOT etc., paren-aware so import blocks are untouched) + repoint the
patches to ldc.<owner>. REPO_ROOT depth bumped parents[2]->[3] for the extra
package level; dead yaml/ac_registry guard import dropped from non-using submodules.

## Verification
- pr_preview: 237 tooling tests pass (no more network hangs from unpatched calls).
- lint_doc: 102 tests pass; the live linter (tools/lint_doc_consistency.py) runs clean.
- Full collection clean (2673); ruff check + format clean on both packages.
