+++ common/ssot/lint_doc_consistency/__init__.py
@@ -1,0 +1,145 @@
+"""Doc/SSOT consistency linter (package)."""
+
+from common.ssot.lint_doc_consistency._base import (
+    AC_PATTERN,
+    AC_REGISTRY,
+    CHECK6_FIXTURE_EXCLUDE,
+    CHECK6_TEST_ROOTS,
+    CI_CD_SSOT,
+    E2E_PRODUCT_TEST_EXCEPTION_PREFIXES,
+    EPIC_DIR,
+    EXCLUDED_DIRS,
+    FRONTEND_SRC,
+    INFRA_REGISTRY,
+    MKDOCS_CONFIG,
+    NO_AC_SCAN_TARGETS,
+    RECONCILIATION_SSOT,
+    REMOVED_ANNOTATION_RE,
+    REPO_ROOT,
+    TDD_SSOT,
+    TEST_FILE_SUFFIXES,
+    TEST_ROOTS,
+    TOTAL_AC_SUMMARY_TOKEN,
+    TRACEABILITY_EXCEPTIONS,
+    VISION_PATH,
+)
+from common.ssot.lint_doc_consistency._types import Violation
+from common.ssot.lint_doc_consistency._parsing import (
+    EPIC_FILE_PATTERN,
+    FRONTEND_RAW_FETCH_ALLOWED_FILES,
+    GENERATED_ANALYSIS_SNAPSHOTS,
+    HTML_ANCHOR_PATTERN,
+    MARKDOWN_CODE_SPAN_PATTERN,
+    MKDOCS_REQUIRED_STATIC_DOCS,
+    PROOF_PLACEMENT_REQUIRED_TOKENS,
+    RAW_FETCH_PATTERN,
+    RECONCILIATION_THRESHOLD_FORBIDDEN_TOKENS,
+    RECONCILIATION_THRESHOLD_REQUIRED_TOKENS,
+    THIN_README_FORBIDDEN_HEADINGS,
+    THIN_README_LIMITS,
+    VISION_ANCHOR_PATTERN,
+    _display_path,
+    _is_excluded_path,
+    _line_is_ac_annotation,
+    _required_mkdocs_nav_docs,
+    collect_ac_refs_in_epics,
+    collect_ac_refs_in_tests,
+    collect_mkdocs_nav_docs,
+    discover_no_ac_test_files,
+    is_deprecated,
+    is_stub,
+    list_epic_files,
+    load_registry_acs,
+    load_traceability_exception_paths,
+    parse_epic_anchor,
+    parse_vision_anchors,
+)
+from common.ssot.lint_doc_consistency._checks import (
+    BACKEND_COVERAGE_COPY_RE,
+    check_code_owned_coverage_threshold_doc,
+    check_epic_anchors,
+    check_epic_to_registry,
+    check_frontend_raw_fetch_usage,
+    check_generated_analysis_snapshots_absent,
+    check_mkdocs_nav_coverage,
+    check_module_readmes_are_thin,
+    check_no_ac_test_exceptions,
+    check_no_e2e_product_test_exceptions,
+    check_orphan_vision_anchors,
+    check_proof_placement_policy,
+    check_reconciliation_thresholds_are_code_owned,
+    check_registry_to_epic,
+    check_registry_to_tests,
+    check_test_id_epic_alignment,
+)
+from common.ssot.lint_doc_consistency.cli import main
+
+__all__ = [
+    "AC_PATTERN",
+    "AC_REGISTRY",
+    "BACKEND_COVERAGE_COPY_RE",
+    "CHECK6_FIXTURE_EXCLUDE",
+    "CHECK6_TEST_ROOTS",
+    "CI_CD_SSOT",
+    "E2E_PRODUCT_TEST_EXCEPTION_PREFIXES",
+    "EPIC_DIR",
+    "EPIC_FILE_PATTERN",
+    "EXCLUDED_DIRS",
+    "FRONTEND_RAW_FETCH_ALLOWED_FILES",
+    "FRONTEND_SRC",
+    "GENERATED_ANALYSIS_SNAPSHOTS",
+    "HTML_ANCHOR_PATTERN",
+    "INFRA_REGISTRY",
+    "MARKDOWN_CODE_SPAN_PATTERN",
+    "MKDOCS_CONFIG",
+    "MKDOCS_REQUIRED_STATIC_DOCS",
+    "NO_AC_SCAN_TARGETS",
+    "PROOF_PLACEMENT_REQUIRED_TOKENS",
+    "RAW_FETCH_PATTERN",
+    "RECONCILIATION_SSOT",
+    "RECONCILIATION_THRESHOLD_FORBIDDEN_TOKENS",
+    "RECONCILIATION_THRESHOLD_REQUIRED_TOKENS",
+    "REMOVED_ANNOTATION_RE",
+    "REPO_ROOT",
+    "TDD_SSOT",
+    "TEST_FILE_SUFFIXES",
+    "TEST_ROOTS",
+    "THIN_README_FORBIDDEN_HEADINGS",
+    "THIN_README_LIMITS",
+    "TOTAL_AC_SUMMARY_TOKEN",
+    "TRACEABILITY_EXCEPTIONS",
+    "VISION_ANCHOR_PATTERN",
+    "VISION_PATH",
+    "Violation",
+    "_display_path",
+    "_is_excluded_path",
+    "_line_is_ac_annotation",
+    "_required_mkdocs_nav_docs",
+    "check_code_owned_coverage_threshold_doc",
+    "check_epic_anchors",
+    "check_epic_to_registry",
+    "check_frontend_raw_fetch_usage",
+    "check_generated_analysis_snapshots_absent",
+    "check_mkdocs_nav_coverage",
+    "check_module_readmes_are_thin",
+    "check_no_ac_test_exceptions",
+    "check_no_e2e_product_test_exceptions",
+    "check_orphan_vision_anchors",
+    "check_proof_placement_policy",
+    "check_reconciliation_thresholds_are_code_owned",
+    "check_registry_to_epic",
+    "check_registry_to_tests",
+    "check_test_id_epic_alignment",
+    "collect_ac_refs_in_epics",
+    "collect_ac_refs_in_tests",
+    "collect_mkdocs_nav_docs",
+    "discover_no_ac_test_files",
+    "is_deprecated",
+    "is_stub",
+    "list_epic_files",
+    "load_registry_acs",
+    "load_traceability_exception_paths",
+    "main",
+    "parse_epic_anchor",
+    "parse_vision_anchors",
+]
