+++ common/ssot/lint_doc_consistency/_types.py
@@ -1,0 +1,10 @@
+"""Violation record."""
+
+from __future__ import annotations
+
+from typing import NamedTuple
+
+
+class Violation(NamedTuple):
+    check: str
+    message: str
