+++ tests/tooling/test_cleanup_pr_preview_resources.py
@@ -13,7 +13,10 @@ def test_AC8_13_38_legacy_cleanup_entrypoints_are_removed() -> None:
 
 
 def test_AC8_13_38_pr_preview_lifecycle_has_no_host_hygiene_commands() -> None:
-    module = (ROOT / "tools/_lib/dev/pr_preview_lifecycle.py").read_text()
+    module = "\n# <<< file-boundary >>>\n".join(
+        p.read_text()
+        for p in sorted((ROOT / "tools/_lib/dev/pr_preview_lifecycle").rglob("*.py"))
+    )
 
     assert "docker builder prune" not in module
     assert "docker image prune" not in module
