--- tools/_lib/dev/pr_preview_lifecycle.py
@@ -1,1623 +1,0 @@
-#!/usr/bin/env python3
-"""Own PR preview create/update/deploy, close cleanup, and reconciliation."""
-
-from __future__ import annotations
-
-import argparse
-import json
-import os
-import re
-import subprocess
-import sys
-import tempfile
-import time
-from collections.abc import Callable
-from dataclasses import dataclass
-from pathlib import Path
-from urllib.parse import quote
-
-try:
-    sys.stdout.reconfigure(line_buffering=True)
-    sys.stderr.reconfigure(line_buffering=True)
-except AttributeError:
-    pass
-
-ALLOWLIST_ENV_KEYS = (
-    "IMAGE_TAG",
-    "GIT_COMMIT_SHA",
-    "IAC_CONFIG_HASH",
-    "ENV_SUFFIX",
-    "ENV_DOMAIN_SUFFIX",
-    "NETWORK_SUFFIX",
-    "NEXT_PUBLIC_API_URL",
-    "DB_HOST",
-    "S3_HOST",
-    "COMPOSE_PROFILES",
-)
-
-# Keys that the platform (Dokploy / Vault-agent) injects or owns in the
-# effective compose env. They legitimately appear in the effective env without
-# being part of the deterministic requested preview env, so reconciliation must
-# not treat them as stale drift (issue #758).
-PLATFORM_MANAGED_ENV_KEYS = (
-    "COMPOSE_PROJECT_NAME",
-    "VAULT_APP_TOKEN",
-)
-PLATFORM_MANAGED_ENV_PREFIXES = (
-    "VAULT_",
-    "DOKPLOY_",
-)
-
-
-def is_platform_managed_env_key(key: str) -> bool:
-    return key in PLATFORM_MANAGED_ENV_KEYS or key.startswith(
-        PLATFORM_MANAGED_ENV_PREFIXES
-    )
-
-
-COMPOSE_SUMMARY_KEYS = (
-    "composeId",
-    "name",
-    "sourceType",
-    "repository",
-    "owner",
-    "branch",
-    "composePath",
-    "command",
-    "environmentId",
-    "composeStatus",
-    "status",
-)
-DEPLOYMENT_DIAGNOSTIC_KEYS = (
-    "message",
-    "error",
-    "errorMessage",
-    "statusMessage",
-    "statusReason",
-    "reason",
-    "description",
-    "logPath",
-)
-SENSITIVE_ASSIGNMENT_PATTERN = re.compile(
-    r"(?i)\b("
-    r"[A-Z0-9_]*(?:TOKEN|SECRET|PASSWORD|PASSWD|API[_-]?KEY|AUTHORIZATION|COOKIE|"
-    r"DATABASE_URL|REFRESH)[A-Z0-9_]*\s*[:=]\s*"
-    r")([^,\s]+)"
-)
-AUTH_VALUE_PATTERN = re.compile(r"(?i)\b(Bearer|Basic)\s+[A-Za-z0-9._~+/\-]+=*")
-VAULT_TOKEN_PATTERN = re.compile(r"\bhvs\.[A-Za-z0-9._-]+")
-
-DEPLOYMENT_READY_FOR_READINESS_STATUSES = {"done"}
-PR_PREVIEW_COMPOSE_PATH = "docker-compose.pr-preview.yml"
-PR_PREVIEW_COMPOSE_TYPE = "docker-compose"
-PR_PREVIEW_SOURCE_TYPE = "github"
-PR_PREVIEW_REPOSITORY = "finance_report"
-PR_PREVIEW_OWNER = "wangzitian0"
-PR_PREVIEW_NEW_DEPLOYMENT_TIMEOUT_SECONDS = 120
-PR_PREVIEW_NEW_DEPLOYMENT_TIMEOUT_SECONDS_ENV = (
-    "PR_PREVIEW_NEW_DEPLOYMENT_TIMEOUT_SECONDS"
-)
-DOKPLOY_API_CONNECT_TIMEOUT_SECONDS = 10
-DOKPLOY_API_MAX_TIME_SECONDS = 20
-PR_PREVIEW_CONTEXT_ENV = "PR_PREVIEW_CONTEXT_PATH"
-
-
-@dataclass(frozen=True)
-class DokployConfig:
-    api_url: str
-    api_key: str
-
-
-class DokployDeploymentDidNotStart(RuntimeError):
-    """Dokploy accepted a deploy request but did not create a new deployment."""
-
-
-class DokployNoNewDeploymentRecord(DokployDeploymentDidNotStart):
-    """Dokploy never created a new deployment record for the requested rollout.
-
-    Distinct from a generic "did not start": the compose can report
-    ``composeStatus=done`` against *stale* deployment records, which previously
-    let the waiter false-green and waste the readiness window probing a route
-    for a SHA that never rolled out (issue #756). This subclass classifies that
-    failure domain (``dokploy-worker-or-deployment-record``) while still being
-    caught by the existing ``DokployDeploymentDidNotStart`` retry handlers.
-    """
-
-
-class DokployDeploymentFailed(RuntimeError):
-    """Dokploy created a deployment record but the rollout failed."""
-
-
-class DokployRequestError(RuntimeError):
-    """Dokploy API request failed after applying the bounded retry policy."""
-
-
-def run_command(
-    cmd: list[str],
-    *,
-    input_text: str | None = None,
-    check: bool = True,
-) -> subprocess.CompletedProcess[str]:
-    return subprocess.run(
-        cmd,
-        input=input_text,
-        capture_output=True,
-        text=True,
-        check=check,
-    )
-
-
-def parse_positive_int_env(var_name: str, default: int) -> int:
-    """Return a positive integer from env, otherwise fallback to default."""
-
-    value = os.environ.get(var_name)
-    if value is None or value == "":
-        return default
-    try:
-        parsed = int(value)
-    except ValueError:
-        print(
-            f"Invalid value for {var_name}; expected integer. "
-            f"falling back to {default}."
-        )
-        return default
-    if parsed <= 0:
-        print(
-            f"Invalid value for {var_name}; expected positive integer. "
-            f"falling back to {default}."
-        )
-        return default
-    return parsed
-
-
-def parse_env(env_text: str) -> dict[str, str]:
-    values: dict[str, str] = {}
-    for line in env_text.splitlines():
-        if not line or line.lstrip().startswith("#") or "=" not in line:
-            continue
-        key, value = line.split("=", 1)
-        values[key.strip()] = value
-    return values
-
-
-def render_env(env: dict[str, str]) -> str:
-    return "\n".join(f"{key}={value}" for key, value in env.items()) + "\n"
-
-
-def allowlisted_env_matches(expected: dict[str, str], actual_env_text: str) -> bool:
-    actual = parse_env(actual_env_text)
-    return all(
-        expected.get(key, "") == actual.get(key, "") for key in ALLOWLIST_ENV_KEYS
-    )
-
-
-def render_allowlisted_env_diff(expected: dict[str, str], actual_env_text: str) -> str:
-    actual = parse_env(actual_env_text)
-    lines = ["Effective deploy env diff"]
-    for key in ALLOWLIST_ENV_KEYS:
-        expected_value = expected.get(key, "")
-        actual_value = actual.get(key, "")
-        if expected_value == actual_value:
-            lines.append(f"{key}: match")
-        else:
-            lines.append(f"{key}: expected={expected_value} actual={actual_value}")
-    lines.append(
-        f"result: {'match' if allowlisted_env_matches(expected, actual_env_text) else 'mismatch'}"
-    )
-    lines.append("raw_env_printed: false")
-    return "\n".join(lines)
-
-
-def env_reconciliation_divergence(
-    requested: dict[str, str], actual_env_text: str
-) -> list[str]:
-    """Return env keys whose effective remote value diverges from the request.
-
-    Covers the whole requested env (not just the allowlist) plus stale keys that
-    linger in the effective env but were never requested — the non-allowlisted
-    drift class from issue #758. Only key *names* are returned so callers can
-    surface them without leaking secret values.
-    """
-
-    actual = parse_env(actual_env_text)
-    divergent: set[str] = set()
-    # A requested key that the effective env *does* expose must match. We do not
-    # flag requested keys the effective payload omits: Dokploy can return a
-    # partial view, and "incomplete echo" is a separate concern from the stale
-    # non-allowlisted drift this guards against.
-    for key, value in requested.items():
-        if key in actual and actual[key] != value:
-            divergent.add(key)
-    # A key present in the effective env that we never requested and that the
-    # platform does not own is a stale leftover from a prior deploy.
-    for key in actual:
-        if key not in requested and not is_platform_managed_env_key(key):
-            divergent.add(key)
-    return sorted(divergent)
-
-
-def render_env_reconciliation_diff(
-    requested: dict[str, str], actual_env_text: str
-) -> str:
-    """Render a secret-safe reconciliation diagnostic (key names only)."""
-
-    divergent = env_reconciliation_divergence(requested, actual_env_text)
-    actual = parse_env(actual_env_text)
-    lines = ["Effective deploy env reconciliation"]
-    lines.append(f"requested_key_count: {len(requested)}")
-    lines.append(f"effective_key_count: {len(actual)}")
-    for key in divergent:
-        # env_reconciliation_divergence only reports keys that are present in the
-        # effective env: either a stale key we never requested, or a requested
-        # key whose effective value differs. It deliberately does not flag keys
-        # absent from the effective env, so those two are the only reasons here.
-        reason = "stale-unrequested-key" if key not in requested else "value-mismatch"
-        lines.append(f"divergent_key: {key} ({reason})")
-    lines.append(f"result: {'match' if not divergent else 'mismatch'}")
-    lines.append("raw_env_printed: false")
-    return "\n".join(lines)
-
-
-def redact_diagnostic_value(value: object) -> str:
-    rendered = str(value)
-    rendered = AUTH_VALUE_PATTERN.sub(r"\1 <redacted>", rendered)
-    rendered = SENSITIVE_ASSIGNMENT_PATTERN.sub(r"\1<redacted>", rendered)
-    rendered = VAULT_TOKEN_PATTERN.sub("hvs.<redacted>", rendered)
-    return rendered[:300]
-
-
-def render_deployment_diagnostics(deployment: dict[str, object]) -> list[str]:
-    lines: list[str] = []
-    for key in DEPLOYMENT_DIAGNOSTIC_KEYS:
-        value = deployment.get(key)
-        if isinstance(value, str) and value:
-            lines.append(f"latest_deployment_{key}: {redact_diagnostic_value(value)}")
-    return lines
-
-
-def render_compose_summary(data: dict[str, object], *, label: str) -> str:
-    lines = [f"Dokploy compose summary ({label})"]
-    for key in COMPOSE_SUMMARY_KEYS:
-        value = data.get(key)
-        if value is None or value == "":
-            continue
-        lines.append(f"{key}: {str(value)[:160]}")
-    lines.append(f"env_present: {bool(data.get('env'))}")
-    deployments = data.get("deployments")
-    if isinstance(deployments, list):
-        lines.append(f"deployment_count: {len(deployments)}")
-        latest = latest_deployment(deployments)
-        if latest:
-            for key in (
-                "deploymentId",
-                "status",
-                "createdAt",
-                "startedAt",
-                "finishedAt",
-            ):
-                value = latest.get(key)
-                if value:
-                    lines.append(f"latest_deployment_{key}: {str(value)[:160]}")
-            lines.extend(render_deployment_diagnostics(latest))
-    lines.append("raw_compose_printed: false")
-    lines.append("raw_deployment_printed: false")
-    return "\n".join(lines)
-
-
-def _deployment_timestamp(deployment: dict[str, object]) -> str:
-    for key in ("createdAt", "startedAt", "finishedAt"):
-        value = deployment.get(key)
-        if isinstance(value, str) and value:
-            return value
-    return ""
-
-
-def latest_deployment(deployments: object) -> dict[str, object] | None:
-    if not isinstance(deployments, list):
-        return None
-    deployment_dicts = [
-        deployment for deployment in deployments if isinstance(deployment, dict)
-    ]
-    if not deployment_dicts:
-        return None
-    return max(deployment_dicts, key=_deployment_timestamp)
-
-
-def deployment_ids(deployments: object) -> set[str]:
-    if not isinstance(deployments, list):
-        return set()
-    ids: set[str] = set()
-    for deployment in deployments:
-        if not isinstance(deployment, dict):
-            continue
-        deployment_id = deployment.get("deploymentId")
-        if deployment_id:
-            ids.add(str(deployment_id))
-    return ids
-
-
-def deployment_signatures(deployments: object) -> dict[str, tuple[str, str, str, str]]:
-    """Capture fields that change when a deployment record is actively updated."""
-
-    if not isinstance(deployments, list):
-        return {}
-    signatures: dict[str, tuple[str, str, str, str]] = {}
-    for deployment in deployments:
-        if not isinstance(deployment, dict):
-            continue
-        deployment_id = deployment.get("deploymentId")
-        if deployment_id:
-            signatures[str(deployment_id)] = (
-                str(deployment.get("status") or ""),
-                str(deployment.get("createdAt") or ""),
-                str(deployment.get("startedAt") or ""),
-                str(deployment.get("finishedAt") or ""),
-            )
-    return signatures
-
-
-def latest_new_deployment(
-    deployments: object, new_deployment_ids: set[str]
-) -> dict[str, object] | None:
-    if not isinstance(deployments, list):
-        return None
-    candidates = [
-        deployment
-        for deployment in deployments
-        if isinstance(deployment, dict)
-        and str(deployment.get("deploymentId") or "") in new_deployment_ids
-    ]
-    return latest_deployment(candidates)
-
-
-def preview_compose_project(pr_number: int) -> str:
-    return f"finance_report_pr_{pr_number}"
-
-
-def preview_commit_slug(commit_sha: str) -> str:
-    if not commit_sha:
-        raise ValueError("commit_sha is required for PR preview routing")
-    slug = re.sub(r"[^a-z0-9]+", "-", commit_sha.lower()).strip("-")
-    if not slug:
-        raise ValueError("commit_sha did not contain any URL-safe characters")
-    return slug[:12]
-
-
-def preview_env_suffix(pr_number: int, commit_sha: str) -> str:
-    return f"-pr-{pr_number}-{preview_commit_slug(commit_sha)}"
-
-
-def preview_stable_app_url(pr_number: int, internal_domain: str) -> str:
-    return f"https://report-pr-{pr_number}.{internal_domain}"
-
-
-def preview_commit_app_url(
-    pr_number: int, commit_sha: str, internal_domain: str
-) -> str:
-    return (
-        f"https://report{preview_env_suffix(pr_number, commit_sha)}.{internal_domain}"
-    )
-
-
-def preview_app_url(pr_number: int, commit_sha: str, internal_domain: str) -> str:
-    return preview_stable_app_url(pr_number, internal_domain)
-
-
-def preview_compose_command(project_name: str | None = None) -> str:
-    # Build the PR's source on the Dokploy host (GitHub-source deploy) instead of
-    # pulling a CI-pushed image. PR previews never push images; image building +
-    # promotion happens only post-merge (staging-deploy). `--build` rebuilds the
-    # backend/frontend contexts that the preview compose now declares.
-    #
-    # The `-p` project name MUST equal Dokploy's own appName for this compose.
-    # Dokploy's compose.delete tears the stack down with `docker compose down`
-    # scoped to appName; if `up` runs under a different project, delete removes
-    # the Dokploy record but leaves the containers orphaned (no `docker compose
-    # down` ever targets them, and nothing else reaps them). Callers pass the
-    # appName via get_compose_app_name(); the project-less form is only a
-    # create-time placeholder that update_compose_source() always overwrites
-    # before any deploy.
-    project_flag = f"-p {project_name} " if project_name else ""
-    return (
-        f"compose {project_flag}"
-        f"-f {PR_PREVIEW_COMPOSE_PATH} up -d --build --remove-orphans"
-    )
-
-
-def preview_port_offset(pr_number: int, commit_sha: str) -> int:
-    seed = f"{pr_number}:{preview_commit_slug(commit_sha)}"
-    return sum((index + 1) * ord(char) for index, char in enumerate(seed)) % 1000
-
-
-def preview_image_tag(pr_number: int, commit_sha: str) -> str:
-    if not commit_sha:
-        raise ValueError("commit_sha is required for PR preview image tags")
-    return f"pr-{pr_number}-{commit_sha}"
-
-
-def validate_deploy_inputs(args: argparse.Namespace) -> None:
-    required_fields = (
-        "pr_number",
-        "compose_name",
-        "environment_id",
-        "api_url",
-        "api_key",
-        "github_integration_id",
-        "branch",
-        "commit_sha",
-        "registry",
-        "image_prefix",
-        "internal_domain",
-    )
-    missing = [
-        field
-        for field in required_fields
-        if not str(getattr(args, field, "") or "").strip()
-    ]
-    if missing:
-        raise ValueError(
-            "Missing required PR preview deploy inputs: " + ", ".join(sorted(missing))
-        )
-    if int(args.pr_number) <= 0:
-        raise ValueError("PR preview deploy requires a positive pr_number")
-    if "/" not in str(args.image_prefix):
-        raise ValueError("PR preview image_prefix must include the registry namespace")
-    if "." not in str(args.internal_domain):
-        raise ValueError("PR preview internal_domain must be a DNS name")
-
-
-def build_preview_context(
-    args: argparse.Namespace,
-    *,
-    phase: str,
-    compose_id: str = "",
-    error: str = "",
-    mutation_step: str = "",
-    recovery_state: str = "",
-) -> dict[str, str]:
-    app_url = preview_app_url(args.pr_number, args.commit_sha, args.internal_domain)
-    context = {
-        "phase": phase,
-        "pr_number": str(args.pr_number),
-        "compose_name": str(args.compose_name),
-        "compose_id": compose_id,
-        "branch": str(args.branch),
-        "commit_sha": str(args.commit_sha),
-        "expected_sha": str(args.commit_sha),
-        "commit_preview_app_url": preview_commit_app_url(
-            args.pr_number,
-            args.commit_sha,
-            args.internal_domain,
-        ),
-        "preview_commit_slug": preview_commit_slug(str(args.commit_sha)),
-        "image_tag": preview_image_tag(args.pr_number, args.commit_sha),
-        "backend_image": f"{args.registry}/{args.image_prefix}-backend:{preview_image_tag(args.pr_number, args.commit_sha)}",
-        "frontend_image": f"{args.registry}/{args.image_prefix}-frontend:{preview_image_tag(args.pr_number, args.commit_sha)}",
-        "app_url": app_url,
-        "api_health_url": f"{app_url}/api/health",
-        "frontend_version_url": f"{app_url}/frontend-version.json?expected={args.commit_sha}",
-        "dokploy_api_url": str(args.api_url).rstrip("/"),
-        "environment_id": str(args.environment_id),
-        "context_schema": "pr-preview-deploy-v1",
-    }
-    if error:
-        context["error"] = redact_diagnostic_value(error)
-    if mutation_step:
-        context["mutation_step"] = mutation_step
-    if recovery_state:
-        context["recovery_state"] = recovery_state
-    return context
-
-
-def write_preview_context(path: str, context: dict[str, str]) -> None:
-    if not path:
-        return
-    context_path = Path(path)
-    context_path.parent.mkdir(parents=True, exist_ok=True)
-    context_path.write_text(
-        json.dumps(context, indent=2, sort_keys=True) + "\n",
-        encoding="utf-8",
-    )
-    print(f"PR preview deploy context written: {context_path}")
-
-
-def build_preview_env(
-    *,
-    pr_number: int,
-    commit_sha: str,
-    registry: str,
-    image_prefix: str,
-    internal_domain: str,
-) -> dict[str, str]:
-    port_offset = preview_port_offset(pr_number, commit_sha)
-    env_suffix = preview_env_suffix(pr_number, commit_sha)
-    app_url = preview_app_url(pr_number, commit_sha, internal_domain)
-    return {
-        "PR_PREVIEW_PR_NUMBER": str(pr_number),
-        "PR_PREVIEW_COMPOSE_NAME": f"pr-{pr_number}",
-        # Informational label only. The real docker compose project is Dokploy's
-        # appName (set via the `-p` flag in preview_compose_command); this stable
-        # per-PR identifier is NOT wired to COMPOSE_PROJECT_NAME anymore.
-        "PR_PREVIEW_COMPOSE_PROJECT": preview_compose_project(pr_number),
-        "PR_PREVIEW_CREATED_BY": "github-actions",
-        "GIT_COMMIT_SHA": commit_sha,
-        "REGISTRY": registry,
-        "IMAGE_PREFIX": image_prefix,
-        "IMAGE_TAG": preview_image_tag(pr_number, commit_sha),
-        "ENV_SUFFIX": env_suffix,
-        "ENV_DOMAIN_SUFFIX": env_suffix,
-        "NETWORK_SUFFIX": f"-pr-{pr_number}",
-        "INTERNAL_DOMAIN": internal_domain,
-        "TRAEFIK_ENABLE": "true",
-        "NEXT_PUBLIC_API_URL": app_url,
-        "NEXT_PUBLIC_APP_URL": app_url,
-        "DEBUG": "true",
-        "DB_PORTS": f"127.0.0.1:{30000 + port_offset}:5432",
-        "MINIO_API_PORTS": f"127.0.0.1:{32000 + port_offset}:9000",
-        "MINIO_CONSOLE_PORTS": f"127.0.0.1:{33000 + port_offset}:9001",
-        "AI_PROVIDER": "zai",
-        "ZAI_API_KEY": "",
-        "AI_BASE_URL": "https://api.z.ai/api/coding/paas/v4",
-        "AI_CHAT_COMPLETIONS_PATH": "/chat/completions",
-        "AI_LAYOUT_PARSING_PATH": "/layout_parsing",
-        "AI_MODEL_CATALOG_SOURCE": "configured",
-        "PRIMARY_MODEL": "glm-5.1",
-        "OCR_MODEL": "glm-4.6v",
-        "VISION_MODEL": "glm-4.6v",
-        "FALLBACK_MODELS": "glm-5-turbo,glm-5",
-        "AI_JSON_TIMEOUT_SECONDS": "360",
-        "AI_JSON_MAX_TOKENS": "8192",
-        "AI_JSON_DISABLE_THINKING": "true",
-        "COMPOSE_PROFILES": "infra,app",
-        "DB_HOST": f"finance-report-db{env_suffix}",
-        "S3_HOST": f"finance-report-minio{env_suffix}",
-        "S3_ENDPOINT": f"http://finance-report-minio{env_suffix}:9000",
-        "MINIO_ROOT_USER": "minio",
-        "MINIO_ROOT_PASSWORD": "minio_local_secret",
-        "S3_ACCESS_KEY": "minio",
-        "S3_SECRET_KEY": "minio_local_secret",
-        "OTEL_EXPORTER_OTLP_ENDPOINT": "http://platform-signoz-otel-collector:4318",
-        "OTEL_SERVICE_NAME": "finance-report-backend",
-        "OTEL_RESOURCE_ATTRIBUTES": f"deployment.environment=pr-{pr_number}",
-        "API_RATE_LIMIT_REQUESTS": "10000",
-        "REGISTER_RATE_LIMIT_REQUESTS": "10000",
-        "IAC_CONFIG_HASH": f"pr-{pr_number}-{commit_sha}",
-    }
-
-
-def _safe_message(body: str) -> str:
-    try:
-        parsed = json.loads(body or "{}")
-    except json.JSONDecodeError:
-        return "unavailable"
-    for key in ("message", "error", "status"):
-        value = parsed.get(key)
-        if isinstance(value, str) and value:
-            return value[:160]
-    return "unavailable"
-
-
-def dokploy_api_retry_delay_seconds() -> float:
-    raw_value = os.environ.get("DOKPLOY_API_RETRY_DELAY_SECONDS", "2.0")
-    try:
-        retry_delay = float(raw_value)
-    except ValueError:
-        return 2.0
-    return retry_delay if retry_delay >= 0 else 2.0
-
-
-def dokploy_api_call(
-    config: DokployConfig,
-    method: str,
-    endpoint: str,
-    *,
-    payload: dict[str, object] | None = None,
-    expected_status: int = 200,
-) -> str:
-    max_attempts = 4 if method == "GET" else 1
-    retry_delay = dokploy_api_retry_delay_seconds()
-
-    for attempt in range(1, max_attempts + 1):
-        config_file = tempfile.NamedTemporaryFile("w", encoding="utf-8", delete=False)
-        data_file: tempfile.NamedTemporaryFile[str] | None = None
-        config_path = config_file.name
-        data_path = ""
-        escaped_api_key = config.api_key.replace("\\", "\\\\").replace('"', '\\"')
-        config_file.write(f'header = "x-api-key: {escaped_api_key}"\n')
-        config_file.close()
-
-        cmd = [
-            "curl",
-            "-sS",
-            "--connect-timeout",
-            str(DOKPLOY_API_CONNECT_TIMEOUT_SECONDS),
-            "--max-time",
-            str(DOKPLOY_API_MAX_TIME_SECONDS),
-            "--config",
-            config_path,
-            "-w",
-            "\n%{http_code}",
-        ]
-        try:
-            if method == "POST":
-                cmd.extend(["-X", "POST", "-H", "Content-Type: application/json"])
-                if payload is not None:
-                    data_file = tempfile.NamedTemporaryFile(
-                        "w", encoding="utf-8", delete=False
-                    )
-                    data_path = data_file.name
-                    json.dump(payload, data_file)
-                    data_file.close()
-                    cmd.extend(["--data-binary", f"@{data_path}"])
-            cmd.append(f"{config.api_url.rstrip('/')}/{endpoint}")
-
-            result = run_command(cmd, check=False)
-        finally:
-            os.unlink(config_path)
-            if data_path:
-                os.unlink(data_path)
-
-        body, _, status_text = result.stdout.rpartition("\n")
-        try:
-            status = int(status_text.strip())
-        except ValueError:
-            status = expected_status if result.returncode == 0 else 0
-            body = result.stdout
-
-        is_transient = (result.returncode == 28) or (status in (500, 502, 503, 504))
-        if (
-            (result.returncode != 0 or status != expected_status)
-            and (attempt < max_attempts)
-            and is_transient
-        ):
-            print(
-                f"Dokploy API call failed (attempt {attempt}/{max_attempts}): "
-                f"endpoint={endpoint} http_code={status or '000'} "
-                f"returncode={result.returncode}. Retrying in {retry_delay}s...",
-                file=sys.stderr,
-            )
-            time.sleep(retry_delay)
-            continue
-
-        if result.returncode != 0 or status != expected_status:
-            print(f"Dokploy request failed: endpoint={endpoint}", file=sys.stderr)
-            print(f"http_code: {status or '000'}", file=sys.stderr)
-            print(f"safe_message: {_safe_message(body)}", file=sys.stderr)
-            print("raw_body_printed: false", file=sys.stderr)
-            if result.stderr:
-                print(result.stderr.strip(), file=sys.stderr)
-            raise DokployRequestError(f"Dokploy request failed for {endpoint}")
-        return body
-
-    raise DokployRequestError(f"Dokploy request failed for {endpoint}")
-
-
-def find_compose_id_by_name(
-    config: DokployConfig, environment_id: str, compose_name: str
-) -> str | None:
-    body = dokploy_api_call(
-        config,
-        "GET",
-        f"environment.one?environmentId={environment_id}",
-    )
-    data = json.loads(body or "{}")
-    for compose in data.get("compose", []):
-        if compose.get("name") == compose_name:
-            compose_id = compose.get("composeId")
-            return str(compose_id) if compose_id else None
-    return None
-
-
-def create_compose(
-    config: DokployConfig,
-    *,
-    environment_id: str,
-    compose_name: str,
-    pr_number: int,
-    branch: str,
-    github_integration_id: str,
-) -> str:
-    body = dokploy_api_call(
-        config,
-        "POST",
-        "compose.create",
-        payload={
-            "name": compose_name,
-            "description": f"PR #{pr_number}: {compose_name}",
-            "environmentId": environment_id,
-            "composeType": PR_PREVIEW_COMPOSE_TYPE,
-            "sourceType": PR_PREVIEW_SOURCE_TYPE,
-            "repository": PR_PREVIEW_REPOSITORY,
-            "owner": PR_PREVIEW_OWNER,
-            "branch": branch,
-            "composePath": PR_PREVIEW_COMPOSE_PATH,
-            "githubId": github_integration_id,
-            # Placeholder: appName does not exist until Dokploy assigns it during
-            # compose.create. update_compose_source() rewrites this with the
-            # appName-scoped command before any deploy.
-            "command": preview_compose_command(),
-            "autoDeploy": False,
-        },
-    )
-    compose_id = json.loads(body or "{}").get("composeId")
-    if not compose_id:
-        raise RuntimeError("Dokploy compose.create response did not include composeId")
-    print(f"Created compose: {compose_id}")
-    return str(compose_id)
-
-
-def get_or_create_compose(
-    config: DokployConfig,
-    *,
-    environment_id: str,
-    compose_name: str,
-    pr_number: int,
-    branch: str,
-    github_integration_id: str,
-) -> str:
-    compose_id = find_compose_id_by_name(config, environment_id, compose_name)
-    if compose_id:
-        print(f"Found existing compose: {compose_id}")
-        return compose_id
-    return create_compose(
-        config,
-        environment_id=environment_id,
-        compose_name=compose_name,
-        pr_number=pr_number,
-        branch=branch,
-        github_integration_id=github_integration_id,
-    )
-
-
-def get_or_create_compose_with_status(
-    config: DokployConfig,
-    *,
-    environment_id: str,
-    compose_name: str,
-    pr_number: int,
-    branch: str,
-    github_integration_id: str,
-) -> tuple[str, bool]:
-    compose_id = find_compose_id_by_name(config, environment_id, compose_name)
-    if compose_id:
-        print(f"Found existing compose: {compose_id}")
-        return compose_id, True
-    return (
-        create_compose(
-            config,
-            environment_id=environment_id,
-            compose_name=compose_name,
-            pr_number=pr_number,
-            branch=branch,
-            github_integration_id=github_integration_id,
-        ),
-        False,
-    )
-
-
-def update_compose_source(
-    config: DokployConfig,
-    *,
-    compose_id: str,
-    branch: str,
-    github_integration_id: str,
-) -> None:
-    # Scope `up` to Dokploy's own appName so compose.delete (which downs the
-    # stack by appName) reaps every container. See preview_compose_command().
-    app_name = get_compose_app_name(config, compose_id=compose_id)
-    dokploy_api_call(
-        config,
-        "POST",
-        "compose.update",
-        payload={
-            "composeId": compose_id,
-            "composeType": PR_PREVIEW_COMPOSE_TYPE,
-            "sourceType": PR_PREVIEW_SOURCE_TYPE,
-            "repository": PR_PREVIEW_REPOSITORY,
-            "owner": PR_PREVIEW_OWNER,
-            "branch": branch,
-            "composePath": PR_PREVIEW_COMPOSE_PATH,
-            "githubId": github_integration_id,
-            "command": preview_compose_command(app_name),
-            "autoDeploy": False,
-        },
-    )
-    print(f"GitHub preview compose configured for compose: {compose_id}")
-
-
-def get_compose_data(config: DokployConfig, *, compose_id: str) -> dict[str, object]:
-    body = dokploy_api_call(
-        config,
-        "GET",
-        f"compose.one?composeId={compose_id}",
-    )
-    data = json.loads(body or "{}")
-    return data if isinstance(data, dict) else {}
-
-
-def get_compose_app_name(config: DokployConfig, *, compose_id: str) -> str:
-    # Dokploy derives the docker compose project from this appName, and
-    # compose.delete downs the stack by it. Fail loud rather than deploy a
-    # preview whose teardown cannot reap its containers (the orphan-leak bug).
-    app_name = str(get_compose_data(config, compose_id=compose_id).get("appName") or "")
-    if not app_name:
-        raise RuntimeError(
-            f"Dokploy compose {compose_id} has no appName; cannot align the "
-            "compose project name with teardown"
-        )
-    return app_name
-
-
-def print_compose_summary(
-    config: DokployConfig, *, compose_id: str, label: str
-) -> None:
-    print(
-        render_compose_summary(
-            get_compose_data(config, compose_id=compose_id), label=label
-        )
-    )
-
-
-def get_compose_env(config: DokployConfig, *, compose_id: str) -> str:
-    data = get_compose_data(config, compose_id=compose_id)
-    env = data.get("env", "")
-    return str(env) if env else ""
-
-
-def update_compose_env(
-    config: DokployConfig, *, compose_id: str, env: dict[str, str]
-) -> None:
-    expected = {key: env[key] for key in ALLOWLIST_ENV_KEYS}
-    dokploy_api_call(
-        config,
-        "POST",
-        "compose.update",
-        payload={"composeId": compose_id, "env": render_env(env)},
-    )
-    effective_env = get_compose_env(config, compose_id=compose_id)
-    print(render_allowlisted_env_diff(expected, effective_env))
-    # Reconcile the *whole* requested env against the effective remote env so a
-    # stale non-allowlisted key from a prior deploy cannot silently diverge
-    # (issue #758). Diagnostics name keys only and never print values.
-    print(render_env_reconciliation_diff(env, effective_env))
-    if not allowlisted_env_matches(expected, effective_env):
-        raise RuntimeError(
-            "Dokploy effective environment did not match requested deploy env"
-        )
-    if env_reconciliation_divergence(env, effective_env):
-        raise RuntimeError(
-            "Dokploy effective environment did not match requested deploy env "
-            "after full reconciliation (stale or divergent non-allowlisted keys)"
-        )
-    print(f"Environment variables configured for compose: {compose_id}")
-
-
-def capture_compose_state(config: DokployConfig, *, compose_id: str) -> dict[str, str]:
-    """Snapshot the last-known-good compose source command and env.
-
-    Captured *before* a deploy mutates the compose so the failure path can roll
-    the record back instead of leaving a half-updated compose pointing at a SHA
-    that never rolled out (issue #758).
-    """
-
-    data = get_compose_data(config, compose_id=compose_id)
-    return {
-        "command": str(data.get("command") or ""),
-        "env": str(data.get("env") or ""),
-    }
-
-
-def restore_compose_state(
-    config: DokployConfig, *, compose_id: str, state: dict[str, str]
-) -> None:
-    """Roll the compose source command and env back to a captured snapshot."""
-
-    payload: dict[str, object] = {"composeId": compose_id}
-    if state.get("command"):
-        payload["command"] = state["command"]
-    payload["env"] = state.get("env", "")
-    dokploy_api_call(config, "POST", "compose.update", payload=payload)
-    print(f"Rolled compose back to last-known-good source/env: {compose_id}")
-
-
-def configure_preview_compose(
-    config: DokployConfig,
-    *,
-    compose_id: str,
-    args: argparse.Namespace,
-    preview_env: dict[str, str],
-    on_step: Callable[[str], None] | None = None,
-) -> None:
-    # ``on_step`` lets the caller mark which mutation the compose was left at
-    # (source vs env) so a failure can report and recover precisely (issue #758).
-    if on_step is not None:
-        on_step("source")
-    update_compose_source(
-        config,
-        compose_id=compose_id,
-        branch=args.branch,
-        github_integration_id=args.github_integration_id,
-    )
-    print_compose_summary(config, compose_id=compose_id, label="after-source-update")
-    if on_step is not None:
-        on_step("env")
-    update_compose_env(
-        config,
-        compose_id=compose_id,
-        env=preview_env,
-    )
-    print_compose_summary(config, compose_id=compose_id, label="after-env-update")
-
-
-def deploy_compose(
-    config: DokployConfig, *, compose_id: str, force_redeploy: bool = False
-) -> None:
-    endpoint = "compose.redeploy" if force_redeploy else "compose.deploy"
-    body = dokploy_api_call(
-        config,
-        "POST",
-        endpoint,
-        payload={"composeId": compose_id},
-    )
-    action = "Redeployment" if force_redeploy else "Deployment"
-    print(f"{action} triggered for compose: {compose_id}")
-    print(f"dokploy_queue_message: {_safe_message(body)}")
-
-
-def deploy_compose_and_wait_for_rollout(
-    config: DokployConfig,
-    *,
-    compose_id: str,
-    force_redeploy: bool,
-) -> None:
-    compose_data = get_compose_data(config, compose_id=compose_id)
-    previous_deployment_ids = deployment_ids(compose_data.get("deployments"))
-    previous_deployment_signatures = deployment_signatures(
-        compose_data.get("deployments")
-    )
-    deploy_compose(config, compose_id=compose_id, force_redeploy=force_redeploy)
-    print_compose_summary(config, compose_id=compose_id, label="after-deploy-trigger")
-    wait_for_dokploy_deployment_rollout(
-        config,
-        compose_id=compose_id,
-        previous_deployment_ids=previous_deployment_ids,
-        previous_deployment_signatures=previous_deployment_signatures,
-    )
-
-
-def fail_before_readiness_after_missing_record(
-    *,
-    compose_id: str,
-    error: DokployDeploymentDidNotStart,
-) -> None:
-    print(
-        "Dokploy deployment record was not created after the deploy request, "
-        "so readiness will not start. "
-        "platform_failure_domain=dokploy-control-plane-record-missing "
-        f"compose_id={compose_id} reason={redact_diagnostic_value(error)}"
-    )
-    print("raw_compose_printed: false")
-    print("raw_deployment_printed: false")
-
-
-def get_running_deployments_count(config: DokployConfig, project_id: str) -> int:
-    try:
-        body = dokploy_api_call(config, "GET", f"project.one?projectId={project_id}")
-        data = json.loads(body or "{}")
-        running_count = 0
-        for env in data.get("environments", []):
-            for comp in env.get("compose", []):
-                if comp.get("composeStatus") in ("running", "deploying"):
-                    running_count += 1
-        return running_count
-    except Exception as exc:
-        print(f"Warning: Failed to check project deployment status: {exc}")
-    return 0
-
-
-def wait_for_dokploy_deployment_rollout(
-    config: DokployConfig,
-    *,
-    compose_id: str,
-    previous_deployment_ids: set[str] | None = None,
-    previous_deployment_signatures: dict[str, tuple[str, str, str, str]] | None = None,
-    timeout_seconds: int = 900,
-    new_deployment_timeout_seconds: int = 600,
-    interval_seconds: int = 5,
-) -> None:
-    previous_deployment_ids = previous_deployment_ids or set()
-    previous_deployment_signatures = previous_deployment_signatures or {}
-    started_at = time.monotonic()
-    deadline = started_at + timeout_seconds
-    new_deployment_deadline = started_at + min(
-        timeout_seconds,
-        new_deployment_timeout_seconds,
-    )
-    attempt = 0
-    while True:
-        attempt += 1
-        now = time.monotonic()
-        try:
-            data = get_compose_data(config, compose_id=compose_id)
-        except DokployRequestError as exc:
-            if now >= deadline:
-                raise
-            print(
-                "Dokploy rollout probe API failure: "
-                f"attempt={attempt} compose_id={compose_id} "
-                f"error={redact_diagnostic_value(exc)}"
-            )
-            time.sleep(interval_seconds)
-            continue
-        compose_status = str(data.get("composeStatus") or "")
-        deployments = data.get("deployments")
-        deployment_count = len(deployments) if isinstance(deployments, list) else 0
-        current_ids = deployment_ids(deployments)
-        new_deployment_ids = sorted(current_ids - previous_deployment_ids)
-        current_deployment_signatures = deployment_signatures(deployments)
-        existing_deployment_updates = sorted(
-            deployment_id
-            for deployment_id in previous_deployment_ids
-            if (
-                deployment_id in current_deployment_signatures
-                and deployment_id in previous_deployment_signatures
-                and current_deployment_signatures[deployment_id]
-                != previous_deployment_signatures[deployment_id]
-            )
-        )
-        if attempt == 1 or attempt % 6 == 0 or compose_status != "idle":
-            print(
-                render_compose_summary(
-                    data, label=f"deployment-rollout-attempt-{attempt}"
-                )
-            )
-        if compose_status == "error":
-            print(
-                render_compose_summary(data, label=f"compose-error-attempt-{attempt}")
-            )
-            if new_deployment_ids:
-                latest = (
-                    latest_new_deployment(deployments, set(new_deployment_ids)) or {}
-                )
-                latest_id = str(latest.get("deploymentId") or new_deployment_ids[-1])
-                raise DokployDeploymentFailed(
-                    "Dokploy compose entered error status before readiness polling: "
-                    f"compose_id={compose_id} deployment_id={latest_id}"
-                )
-            print(
-                "Dokploy compose still reports a stale error before the queued "
-                "redeploy has created a new deployment record; continuing rollout poll."
-            )
-        print(
-            f"Dokploy rollout probe: attempt={attempt} "
-            f"compose_id={compose_id} "
-            f"composeStatus={compose_status or 'unknown'} "
-            f"deployment_count={deployment_count} "
-            f"new_deployment_ids={','.join(new_deployment_ids) or 'none'}"
-        )
-        if (
-            not new_deployment_ids
-            and existing_deployment_updates
-            and compose_status in DEPLOYMENT_READY_FOR_READINESS_STATUSES
-        ):
-            existing_latest = (
-                latest_new_deployment(deployments, set(existing_deployment_updates))
-                or {}
-            )
-            existing_latest_id = str(
-                existing_latest.get("deploymentId") or existing_deployment_updates[-1]
-            )
-            existing_latest_status = str(existing_latest.get("status") or "unknown")
-            print(
-                "Dokploy rollout observed as existing deployment record update: "
-                f"compose_id={compose_id} "
-                f"existing_deployment_ids={','.join(existing_deployment_updates)} "
-                f"latest_deployment_id={existing_latest_id} "
-                f"latest_deployment_status={existing_latest_status}"
-            )
-            if existing_latest_status == "error":
-                print(
-                    render_compose_summary(
-                        data,
-                        label=f"existing-deployment-error-attempt-{attempt}",
-                    )
-                )
-                raise DokployDeploymentFailed(
-                    "Dokploy deployment failed before readiness polling: "
-                    f"compose_id={compose_id} deployment_id={existing_latest_id}"
-                )
-            if existing_latest_status in DEPLOYMENT_READY_FOR_READINESS_STATUSES:
-                return
-        if new_deployment_ids:
-            latest = latest_new_deployment(deployments, set(new_deployment_ids)) or {}
-            latest_id = str(latest.get("deploymentId") or new_deployment_ids[-1])
-            latest_status = str(latest.get("status") or "unknown")
-            if latest_status == "error":
-                print(
-                    render_compose_summary(
-                        data,
-                        label=f"deployment-error-attempt-{attempt}",
-                    )
-                )
-                raise DokployDeploymentFailed(
-                    "Dokploy deployment failed before readiness polling: "
-                    f"compose_id={compose_id} deployment_id={latest_id}"
-                )
-            print(
-                f"Dokploy deployment observed: compose_id={compose_id} "
-                f"composeStatus={compose_status or 'unknown'} "
-                f"deployment_count={deployment_count} "
-                f"new_deployment_ids={','.join(new_deployment_ids)} "
-                f"latest_deployment_id={latest_id} "
-                f"latest_deployment_status={latest_status}"
-            )
-            if latest_status in DEPLOYMENT_READY_FOR_READINESS_STATUSES:
-                return
-        if (
-            not new_deployment_ids
-            and current_ids
-            and compose_status in DEPLOYMENT_READY_FOR_READINESS_STATUSES
-            and now >= new_deployment_deadline
-        ):
-            print(
-                "Dokploy did not create a new deployment record, but compose "
-                "is done with existing deployment records; failing before "
-                "application readiness instead of probing the route for a SHA "
-                "that never rolled out. platform_failure_domain="
-                "dokploy-worker-or-deployment-record"
-            )
-            raise DokployNoNewDeploymentRecord(
-                "Dokploy compose is done with existing deployment records but "
-                "did not create a new deployment record for this rollout before "
-                f"readiness polling: compose_id={compose_id} "
-                f"composeStatus={compose_status or 'unknown'} "
-                f"deployment_count={deployment_count} "
-                "platform_failure_domain=dokploy-worker-or-deployment-record"
-            )
-        if not new_deployment_ids and now >= new_deployment_deadline:
-            project_id = (
-                data.get("environment", {}).get("projectId")
-                if isinstance(data, dict)
-                else None
-            )
-            if (
-                project_id
-                and get_running_deployments_count(config, project_id) > 0
-                and now < deadline
-            ):
-                print(
-                    "Warning: Dokploy is currently busy with other deployments. "
-                    "Extending the new deployment timeout deadline."
-                )
-                new_deployment_deadline = min(now + 60.0, deadline)
-                continue
-            raise DokployNoNewDeploymentRecord(
-                "Dokploy deployment did not create a new deployment before "
-                f"readiness polling: compose_id={compose_id} "
-                f"composeStatus={compose_status or 'unknown'} "
-                f"deployment_count={deployment_count} "
-                "platform_failure_domain=dokploy-worker-or-deployment-record"
-            )
-        if now >= deadline:
-            raise RuntimeError(
-                "Dokploy deployment did not reach done before readiness "
-                f"polling: compose_id={compose_id} composeStatus={compose_status or 'unknown'} "
-                f"deployment_count={deployment_count} "
-                f"new_deployment_ids={','.join(new_deployment_ids)}"
-            )
-        time.sleep(interval_seconds)
-
-
-def delete_compose(config: DokployConfig, *, compose_id: str) -> None:
-    dokploy_api_call(
-        config,
-        "POST",
-        "compose.delete",
-        payload={"composeId": compose_id},
-    )
-    print(f"Compose deleted: {compose_id}")
-
-
-def deploy_action(args: argparse.Namespace) -> int:
-    validate_deploy_inputs(args)
-    config = DokployConfig(api_url=args.api_url, api_key=args.api_key)
-    context_path = os.environ.get(PR_PREVIEW_CONTEXT_ENV, "")
-    compose_id = ""
-    # Tracks which mutation the compose was left at and the last-known-good
-    # snapshot to roll back to on failure (issue #758). "step" advances as the
-    # deploy mutates source -> env -> deploy -> rollout. "good" is None until we
-    # have a snapshot worth restoring (only existing composes have one).
-    mutation: dict[str, object] = {"step": "preflight", "good": None}
-
-    def record_context(
-        phase: str,
-        *,
-        error: str = "",
-        mutation_step: str = "",
-        recovery_state: str = "",
-    ) -> None:
-        write_preview_context(
-            context_path,
-            build_preview_context(
-                args,
-                phase=phase,
-                compose_id=compose_id,
-                error=error,
-                mutation_step=mutation_step,
-                recovery_state=recovery_state,
-            ),
-        )
-
-    record_context("preflight")
-    try:
-        compose_id, existing_compose = get_or_create_compose_with_status(
-            config,
-            environment_id=args.environment_id,
-            compose_name=args.compose_name,
-            pr_number=args.pr_number,
-            branch=args.branch,
-            github_integration_id=args.github_integration_id,
-        )
-        record_context("compose-resolved")
-        existing_compose_data = (
-            get_compose_data(config, compose_id=compose_id) if existing_compose else {}
-        )
-        if (
-            existing_compose
-            and str(existing_compose_data.get("composeStatus") or "") == "idle"
-            and not deployment_ids(existing_compose_data.get("deployments"))
-        ):
-            print(
-                "Existing preview compose has no deployment records; "
-                "recreating before deploy"
-            )
-            delete_compose(config, compose_id=compose_id)
-            compose_id = create_compose(
-                config,
-                environment_id=args.environment_id,
-                compose_name=args.compose_name,
-                pr_number=args.pr_number,
-                branch=args.branch,
-                github_integration_id=args.github_integration_id,
-            )
-            existing_compose = False
-            record_context("compose-recreated")
-
-        preview_env = build_preview_env(
-            pr_number=args.pr_number,
-            commit_sha=args.commit_sha,
-            registry=args.registry,
-            image_prefix=args.image_prefix,
-            internal_domain=args.internal_domain,
-        )
-
-        def configure_current_compose() -> None:
-            # Snapshot the last-known-good source/env of an *existing* compose
-            # before mutating it, so a later rollout failure can roll back rather
-            # than leave a half-updated compose (issue #758). A freshly-created
-            # compose has no good state to restore, so leave the snapshot unset
-            # and fall back to the marked safe-to-reconcile path.
-            if existing_compose and mutation["good"] is None:
-                mutation["good"] = capture_compose_state(config, compose_id=compose_id)
-
-            def mark_step(step: str) -> None:
-                mutation["step"] = step
-
-            configure_preview_compose(
-                config,
-                compose_id=compose_id,
-                args=args,
-                preview_env=preview_env,
-                on_step=mark_step,
-            )
-
-        def trigger_and_wait(*, force_redeploy: bool) -> None:
-            compose_data = get_compose_data(config, compose_id=compose_id)
-            previous_deployment_ids = deployment_ids(compose_data.get("deployments"))
-            previous_deployment_signatures = deployment_signatures(
-                compose_data.get("deployments")
-            )
-            mutation["step"] = "deploy"
-            deploy_compose(config, compose_id=compose_id, force_redeploy=force_redeploy)
-            print_compose_summary(
-                config,
-                compose_id=compose_id,
-                label="after-redeploy-trigger"
-                if force_redeploy
-                else "after-deploy-trigger",
-            )
-            record_context(
-                "redeploy-triggered" if force_redeploy else "deploy-triggered"
-            )
-            mutation["step"] = "rollout"
-            wait_for_dokploy_deployment_rollout(
-                config,
-                compose_id=compose_id,
-                previous_deployment_ids=previous_deployment_ids,
-                previous_deployment_signatures=previous_deployment_signatures,
-                new_deployment_timeout_seconds=parse_positive_int_env(
-                    PR_PREVIEW_NEW_DEPLOYMENT_TIMEOUT_SECONDS_ENV,
-                    PR_PREVIEW_NEW_DEPLOYMENT_TIMEOUT_SECONDS,
-                ),
-            )
-
-        def recreate_compose_before_retry() -> None:
-            nonlocal compose_id, existing_compose
-            # The compose is being torn down and recreated fresh; there is no
-            # prior good state to roll back to anymore.
-            mutation["good"] = None
-            delete_compose(config, compose_id=compose_id)
-            compose_id = create_compose(
-                config,
-                environment_id=args.environment_id,
-                compose_name=args.compose_name,
-                pr_number=args.pr_number,
-                branch=args.branch,
-                github_integration_id=args.github_integration_id,
-            )
-            existing_compose = False
-            record_context("compose-recreated")
-            configure_current_compose()
-
-        configure_current_compose()
-
-        try:
-            trigger_and_wait(force_redeploy=existing_compose)
-        except DokployDeploymentDidNotStart:
-            if existing_compose:
-                print(
-                    "Existing PR preview compose did not complete a new Dokploy "
-                    "rollout; recreating compose before retry."
-                )
-                recreate_compose_before_retry()
-                try:
-                    trigger_and_wait(force_redeploy=False)
-                except DokployDeploymentDidNotStart as retry_error:
-                    fail_before_readiness_after_missing_record(
-                        compose_id=compose_id,
-                        error=retry_error,
-                    )
-                    record_context("failed", error=str(retry_error))
-                    raise
-            else:
-                print(
-                    "Initial Dokploy deploy did not create a deployment record; "
-                    "retrying with compose.redeploy"
-                )
-                try:
-                    trigger_and_wait(force_redeploy=True)
-                except DokployDeploymentDidNotStart:
-                    print(
-                        "New PR preview compose still did not create a Dokploy "
-                        "deployment record after redeploy; recreating compose "
-                        "before final retry."
-                    )
-                    recreate_compose_before_retry()
-                    try:
-                        trigger_and_wait(force_redeploy=False)
-                    except DokployDeploymentDidNotStart as recreate_error:
-                        fail_before_readiness_after_missing_record(
-                            compose_id=compose_id,
-                            error=recreate_error,
-                        )
-                        record_context("failed", error=str(recreate_error))
-                        raise
-        except DokployDeploymentFailed:
-            if not existing_compose:
-                raise
-            print(
-                "Existing PR preview compose did not complete a new Dokploy "
-                "rollout; recreating compose before retry."
-            )
-            recreate_compose_before_retry()
-            trigger_and_wait(force_redeploy=False)
-
-        record_context("rollout-ready")
-
-        if github_output := os.environ.get("GITHUB_OUTPUT"):
-            with open(github_output, "a", encoding="utf-8") as output:
-                output.write(f"compose_id={compose_id}\n")
-                output.write(
-                    f"app_url={preview_app_url(args.pr_number, args.commit_sha, args.internal_domain)}\n"
-                )
-        else:
-            print(f"compose_id={compose_id}")
-            print(
-                f"app_url={preview_app_url(args.pr_number, args.commit_sha, args.internal_domain)}"
-            )
-        return 0
-    except Exception as exc:
-        # Issue #758: never leave a silent half-update. If we mutated an existing
-        # compose, roll it back to the captured last-known-good source/env;
-        # otherwise (a freshly-created or already-recreated compose) explicitly
-        # mark the record safe-to-reconcile so a retry/reconcile knows the
-        # compose is not a trustworthy running state. Always record which
-        # mutation step the compose was left at.
-        mutation_step = str(mutation.get("step") or "preflight")
-        recovery_state = "marked-safe-to-reconcile"
-        good_state = mutation.get("good")
-        if isinstance(good_state, dict) and compose_id:
-            try:
-                restore_compose_state(config, compose_id=compose_id, state=good_state)
-                recovery_state = "rolled-back"
-            except Exception as rollback_exc:  # pragma: no cover - defensive
-                print(
-                    "WARNING: failed to roll preview compose back to "
-                    f"last-known-good; left safe-to-reconcile: "
-                    f"{type(rollback_exc).__name__}: "
-                    f"{redact_diagnostic_value(rollback_exc)}"
-                )
-        print(
-            "PR preview deploy left compose in a recovery state: "
-            f"compose_id={compose_id or 'none'} "
-            f"mutation_step={mutation_step} recovery_state={recovery_state}"
-        )
-        record_context(
-            "failed",
-            error=f"{type(exc).__name__}: {exc}",
-            mutation_step=mutation_step,
-            recovery_state=recovery_state,
-        )
-        print(
-            "PR preview deploy failed: "
-            f"{type(exc).__name__}: {redact_diagnostic_value(exc)}"
-        )
-        return 1
-
-
-def cleanup_action(args: argparse.Namespace) -> int:
-    try:
-        config = DokployConfig(api_url=args.api_url, api_key=args.api_key)
-        compose_id = args.compose_id or find_compose_id_by_name(
-            config,
-            args.environment_id,
-            args.compose_name,
-        )
-        if compose_id:
-            delete_compose(config, compose_id=compose_id)
-        else:
-            print(f"Compose not found: {args.compose_name}")
-    except DokployRequestError as exc:
-        print(
-            f"WARNING: Cleanup action failed for {args.compose_name} (ignoring): {exc}",
-            file=sys.stderr,
-        )
-    return 0
-
-
-def delete_action(args: argparse.Namespace) -> int:
-    try:
-        config = DokployConfig(api_url=args.api_url, api_key=args.api_key)
-        compose_id = args.compose_id or find_compose_id_by_name(
-            config,
-            args.environment_id,
-            args.compose_name,
-        )
-        if not compose_id:
-            print(f"Compose not found: {args.compose_name}")
-            return 0
-        delete_compose(config, compose_id=compose_id)
-    except DokployRequestError as exc:
-        print(
-            f"WARNING: Delete action failed for {args.compose_name} (ignoring): {exc}",
-            file=sys.stderr,
-        )
-    return 0
-
-
-def parse_open_pr_numbers(output: str) -> set[int]:
-    return {int(line.strip()) for line in output.splitlines() if line.strip()}
-
-
-def list_open_pr_numbers() -> set[int]:
-    result = run_command(
-        [
-            "gh",
-            "pr",
-            "list",
-            "--state",
-            "open",
-            "--limit",
-            "1000",
-            "--json",
-            "number",
-            "--jq",
-            ".[].number",
-        ]
-    )
-    return parse_open_pr_numbers(result.stdout)
-
-
-def parse_preview_pr_from_compose_name(name: str) -> int | None:
-    match = re.fullmatch(r"pr-([1-9][0-9]*)", name)
-    return int(match.group(1)) if match else None
-
-
-def list_preview_composes(config: DokployConfig, environment_id: str) -> dict[int, str]:
-    body = dokploy_api_call(
-        config,
-        "GET",
-        f"environment.one?environmentId={quote(environment_id, safe='')}",
-    )
-    data = json.loads(body or "{}")
-    previews: dict[int, str] = {}
-    for compose in data.get("compose", []):
-        name = str(compose.get("name") or "")
-        pr_number = parse_preview_pr_from_compose_name(name)
-        compose_id = compose.get("composeId")
-        if pr_number is not None and compose_id:
-            previews[pr_number] = str(compose_id)
-    return previews
-
-
-def reconcile_action(args: argparse.Namespace) -> int:
-    open_prs = list_open_pr_numbers()
-    config = DokployConfig(api_url=args.api_url, api_key=args.api_key)
-    preview_composes = list_preview_composes(config, args.environment_id)
-    remote_prs = set(preview_composes)
-    stale_prs = sorted(remote_prs - open_prs)
-    print(f"Open PRs: {sorted(open_prs)}")
-    print(f"Preview PRs in Dokploy: {sorted(remote_prs)}")
-    print(f"Stale preview PRs: {stale_prs}")
-    for pr_number in stale_prs:
-        compose_id = preview_composes[pr_number]
-        if args.dry_run:
-            print(f"[dry-run] Would delete compose for PR #{pr_number}: {compose_id}")
-        else:
-            delete_compose(config, compose_id=compose_id)
-    return 0
-
-
-def main_from_args(args: argparse.Namespace) -> int:
-    if args.action == "deploy":
-        return deploy_action(args)
-    if args.action == "delete":
-        return delete_action(args)
-    if args.action == "cleanup":
-        return cleanup_action(args)
-    if args.action == "reconcile":
-        return reconcile_action(args)
-    raise ValueError(f"Unsupported action: {args.action}")
-
-
-def normalize_dash_prefixed_values(argv: list[str]) -> list[str]:
-    normalized: list[str] = []
-    index = 0
-    value_options = {"--environment-id"}
-    while index < len(argv):
-        current = argv[index]
-        if current in value_options and index + 1 < len(argv):
-            normalized.append(f"{current}={argv[index + 1]}")
-            index += 2
-            continue
-        normalized.append(current)
-        index += 1
-    return normalized
-
-
-def main(argv: list[str] | None = None) -> int:
-    argv = normalize_dash_prefixed_values(
-        list(argv) if argv is not None else sys.argv[1:]
-    )
-    parser = argparse.ArgumentParser(description=__doc__)
-    parser.add_argument(
-        "--action", choices=["deploy", "delete", "cleanup", "reconcile"], required=True
-    )
-    parser.add_argument("--pr-number", type=int, required=True)
-    parser.add_argument("--compose-name", required=True)
-    parser.add_argument("--compose-id", default="")
-    parser.add_argument("--environment-id", required=True)
-    parser.add_argument("--api-url", required=True)
-    parser.add_argument("--api-key", required=True)
-    parser.add_argument("--github-integration-id", default="")
-    parser.add_argument("--branch", default="")
-    parser.add_argument("--commit-sha", default="")
-    parser.add_argument("--registry", default="ghcr.io")
-    parser.add_argument("--image-prefix", default="")
-    parser.add_argument("--internal-domain", default="zitian.party")
-    parser.add_argument("--dry-run", action="store_true")
-    return main_from_args(parser.parse_args(argv))
-
-
-if __name__ == "__main__":
-    sys.exit(main())
