+++ tools/_lib/dev/pr_preview_lifecycle/__init__.py
@@ -1,0 +1,191 @@
+"""PR preview lifecycle (package)."""
+
+import time  # noqa: F401  (kept for monkeypatch of lifecycle.time)
+
+from tools._lib.dev.pr_preview_lifecycle._base import (
+    ALLOWLIST_ENV_KEYS,
+    AUTH_VALUE_PATTERN,
+    COMPOSE_SUMMARY_KEYS,
+    DEPLOYMENT_DIAGNOSTIC_KEYS,
+    DEPLOYMENT_READY_FOR_READINESS_STATUSES,
+    DOKPLOY_API_CONNECT_TIMEOUT_SECONDS,
+    DOKPLOY_API_MAX_TIME_SECONDS,
+    PLATFORM_MANAGED_ENV_KEYS,
+    PLATFORM_MANAGED_ENV_PREFIXES,
+    PR_PREVIEW_COMPOSE_PATH,
+    PR_PREVIEW_COMPOSE_TYPE,
+    PR_PREVIEW_CONTEXT_ENV,
+    PR_PREVIEW_NEW_DEPLOYMENT_TIMEOUT_SECONDS,
+    PR_PREVIEW_NEW_DEPLOYMENT_TIMEOUT_SECONDS_ENV,
+    PR_PREVIEW_OWNER,
+    PR_PREVIEW_REPOSITORY,
+    PR_PREVIEW_SOURCE_TYPE,
+    SENSITIVE_ASSIGNMENT_PATTERN,
+    VAULT_TOKEN_PATTERN,
+    is_platform_managed_env_key,
+)
+from tools._lib.dev.pr_preview_lifecycle._util import (
+    _deployment_timestamp,
+    _safe_message,
+    allowlisted_env_matches,
+    deployment_ids,
+    deployment_signatures,
+    dokploy_api_retry_delay_seconds,
+    env_reconciliation_divergence,
+    latest_deployment,
+    latest_new_deployment,
+    normalize_dash_prefixed_values,
+    parse_env,
+    parse_positive_int_env,
+    redact_diagnostic_value,
+    render_allowlisted_env_diff,
+    render_compose_summary,
+    render_deployment_diagnostics,
+    render_env,
+    render_env_reconciliation_diff,
+    run_command,
+)
+from tools._lib.dev.pr_preview_lifecycle._preview import (
+    build_preview_context,
+    build_preview_env,
+    preview_app_url,
+    preview_commit_app_url,
+    preview_commit_slug,
+    preview_compose_command,
+    preview_compose_project,
+    preview_env_suffix,
+    preview_image_tag,
+    preview_port_offset,
+    preview_stable_app_url,
+    validate_deploy_inputs,
+    write_preview_context,
+)
+from tools._lib.dev.pr_preview_lifecycle._dokploy import (
+    DokployConfig,
+    DokployDeploymentDidNotStart,
+    DokployDeploymentFailed,
+    DokployNoNewDeploymentRecord,
+    DokployRequestError,
+    capture_compose_state,
+    configure_preview_compose,
+    create_compose,
+    delete_compose,
+    deploy_compose,
+    deploy_compose_and_wait_for_rollout,
+    dokploy_api_call,
+    fail_before_readiness_after_missing_record,
+    find_compose_id_by_name,
+    get_compose_app_name,
+    get_compose_data,
+    get_compose_env,
+    get_or_create_compose,
+    get_or_create_compose_with_status,
+    get_running_deployments_count,
+    list_preview_composes,
+    parse_preview_pr_from_compose_name,
+    print_compose_summary,
+    restore_compose_state,
+    update_compose_env,
+    update_compose_source,
+    wait_for_dokploy_deployment_rollout,
+)
+from tools._lib.dev.pr_preview_lifecycle.cli import (
+    cleanup_action,
+    delete_action,
+    deploy_action,
+    list_open_pr_numbers,
+    main,
+    main_from_args,
+    parse_open_pr_numbers,
+    reconcile_action,
+)
+
+__all__ = [
+    "ALLOWLIST_ENV_KEYS",
+    "AUTH_VALUE_PATTERN",
+    "COMPOSE_SUMMARY_KEYS",
+    "DEPLOYMENT_DIAGNOSTIC_KEYS",
+    "DEPLOYMENT_READY_FOR_READINESS_STATUSES",
+    "DOKPLOY_API_CONNECT_TIMEOUT_SECONDS",
+    "DOKPLOY_API_MAX_TIME_SECONDS",
+    "DokployConfig",
+    "DokployDeploymentDidNotStart",
+    "DokployDeploymentFailed",
+    "DokployNoNewDeploymentRecord",
+    "DokployRequestError",
+    "PLATFORM_MANAGED_ENV_KEYS",
+    "PLATFORM_MANAGED_ENV_PREFIXES",
+    "PR_PREVIEW_COMPOSE_PATH",
+    "PR_PREVIEW_COMPOSE_TYPE",
+    "PR_PREVIEW_CONTEXT_ENV",
+    "PR_PREVIEW_NEW_DEPLOYMENT_TIMEOUT_SECONDS",
+    "PR_PREVIEW_NEW_DEPLOYMENT_TIMEOUT_SECONDS_ENV",
+    "PR_PREVIEW_OWNER",
+    "PR_PREVIEW_REPOSITORY",
+    "PR_PREVIEW_SOURCE_TYPE",
+    "SENSITIVE_ASSIGNMENT_PATTERN",
+    "VAULT_TOKEN_PATTERN",
+    "_deployment_timestamp",
+    "_safe_message",
+    "allowlisted_env_matches",
+    "build_preview_context",
+    "build_preview_env",
+    "capture_compose_state",
+    "cleanup_action",
+    "configure_preview_compose",
+    "create_compose",
+    "delete_action",
+    "delete_compose",
+    "deploy_action",
+    "deploy_compose",
+    "deploy_compose_and_wait_for_rollout",
+    "deployment_ids",
+    "deployment_signatures",
+    "dokploy_api_call",
+    "dokploy_api_retry_delay_seconds",
+    "env_reconciliation_divergence",
+    "fail_before_readiness_after_missing_record",
+    "find_compose_id_by_name",
+    "get_compose_app_name",
+    "get_compose_data",
+    "get_compose_env",
+    "get_or_create_compose",
+    "get_or_create_compose_with_status",
+    "get_running_deployments_count",
+    "is_platform_managed_env_key",
+    "latest_deployment",
+    "latest_new_deployment",
+    "list_open_pr_numbers",
+    "list_preview_composes",
+    "main",
+    "main_from_args",
+    "normalize_dash_prefixed_values",
+    "parse_env",
+    "parse_open_pr_numbers",
+    "parse_positive_int_env",
+    "parse_preview_pr_from_compose_name",
+    "preview_app_url",
+    "preview_commit_app_url",
+    "preview_commit_slug",
+    "preview_compose_command",
+    "preview_compose_project",
+    "preview_env_suffix",
+    "preview_image_tag",
+    "preview_port_offset",
+    "preview_stable_app_url",
+    "print_compose_summary",
+    "reconcile_action",
+    "redact_diagnostic_value",
+    "render_allowlisted_env_diff",
+    "render_compose_summary",
+    "render_deployment_diagnostics",
+    "render_env",
+    "render_env_reconciliation_diff",
+    "restore_compose_state",
+    "run_command",
+    "update_compose_env",
+    "update_compose_source",
+    "validate_deploy_inputs",
+    "wait_for_dokploy_deployment_rollout",
+    "write_preview_context",
+]
